+++ src/app/compare/Compare.tsx
@@ -16,16 +16,21 @@ import { Subscriptions } from '../rx-utils';
 import { connect } from 'react-redux';
 import { ReviewsState, getRating } from '../item-review/reducer';
 import { RootState } from '../store/reducers';
+import { CurationsState } from '../curated-rolls/reducer';
 
 interface StoreProps {
   ratings: ReviewsState['ratings'];
+  curationEnabled: boolean;
+  curations: CurationsState['curations'];
 }
 
 type Props = StoreProps;
 
 function mapStateToProps(state: RootState): StoreProps {
   return {
-    ratings: state.reviews.ratings
+    ratings: state.reviews.ratings,
+    curationEnabled: state.curations.curationEnabled,
+    curations: state.curations.curations
   };
 }
 
@@ -174,6 +179,8 @@ class Compare extends React.Component<Props, State> {
                 setHighlight={this.setHighlight}
                 $scope={this.$scope}
                 highlight={highlight}
+                curationEnabled={this.props.curationEnabled}
+                inventoryCuratedRoll={this.props.curations[item.id]}
               />
             ))}
           </div>